# jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Solution to combine municipality services under single API.",
    "version" : "0.1.0",
    "title" : "Kunta API"
  },
  "host" : "demo.kuntaapi.fi",
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "paths" : {
...
    "/printableFormServiceChannels" : {
      "get" : {
        "tags" : [ "ServiceChannels", "PrintableFormServiceChannels" ],
        "summary" : "Lists printable form service channels",
        "description" : "Lists printable form service channels",
        "operationId" : "listPrintableFormServiceChannels",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns array of printable form service channels",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PrintableFormServiceChannel"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/printableFormServiceChannels/{printableFormServiceChannelId}" : {
      "get" : {
        "tags" : [ "ServiceChannels", "PrintableFormServiceChannels" ],
        "summary" : "Finds a printable form service channel",
        "description" : "Finds a printable form service channel",
        "operationId" : "findPrintableFormServiceChannel",
        "parameters" : [ {
          "name" : "printableFormServiceChannelId",
          "in" : "path",
          "description" : "Printable form service channel id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns printable form service channel",
            "schema" : {
              "$ref" : "#/definitions/PrintableFormServiceChannel"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/serviceLocationServiceChannels" : {
      "get" : {
        "tags" : [ "ServiceChannels", "ServiceLocationServiceChannels" ],
        "summary" : "Lists service location service channels",
        "description" : "Lists service location service channels",
        "operationId" : "listServiceLocationServiceChannels",
        "parameters" : [ {
          "name" : "organizationId",
          "in" : "query",
          "description" : "Organization id",
          "required" : false,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "search",
          "in" : "query",
          "description" : "Search service location channels by free-text query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns array of service location service channels",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ServiceLocationServiceChannel"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/serviceLocationServiceChannels/{serviceLocationServiceChannelId}" : {
      "get" : {
        "tags" : [ "ServiceChannels", "ServiceLocationServiceChannels" ],
        "summary" : "Finds a service location service by id",
        "description" : "Finds a service location service by id",
        "operationId" : "findServiceLocationServiceChannel",
        "parameters" : [ {
          "name" : "serviceLocationServiceChannelId",
          "in" : "path",
          "description" : "serviceLocationChannel id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a service location service channel",
            "schema" : {
              "$ref" : "#/definitions/ServiceLocationServiceChannel"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
    "/services/{serviceId}" : {
      "get" : {
        "tags" : [ "Services" ],
        "summary" : "Finds a service by id",
        "description" : "Finds service",
        "operationId" : "findService",
        "parameters" : [ {
          "name" : "serviceId",
          "in" : "path",
          "description" : "Service id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a service",
            "schema" : {
              "$ref" : "#/definitions/Service"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Services" ],
        "summary" : "Updates service",
        "description" : "Updates service",
        "operationId" : "updateService",
        "parameters" : [ {
          "name" : "serviceId",
          "in" : "path",
          "description" : "Organization service id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Payload",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Service"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns an updated service",
            "schema" : {
              "$ref" : "#/definitions/Service"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "basicAuth" : {
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "Service" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "type" : {
          "type" : "string"
        },
        "statutoryDescriptionId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "chargeType" : {
          "type" : "string"
        },
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "areaType" : {
          "type" : "string",
          "description" : "Area type (WholeCountry, WholeCountryExceptAlandIslands, AreaType)."
        },
        "areas" : {
          "type" : "array",
          "description" : "List of service areas.",
          "items" : {
            "$ref" : "#/definitions/Area"
          }
        },
        "descriptions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "languages" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "serviceClasses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "ontologyTerms" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "targetGroups" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "lifeEvents" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "industrialClasses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "legislation" : {
          "type" : "array",
          "description" : "List of laws related to the service.",
          "items" : {
            "$ref" : "#/definitions/Law"
          }
        },
        "keywords" : {
          "type" : "array",
          "description" : "List of localized service keywords.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "requirements" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "publishingStatus" : {
          "type" : "string",
          "description" : "Publishing status. Possible values are: Draft, Published, Deleted, Modified or OldPublished."
        },
        "organizations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ServiceOrganization"
          }
        },
        "electronicServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "phoneServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "printableFormServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "serviceLocationServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "webPageServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      }
    },
...
    "PrintableFormServiceChannel" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Identifier for the service channel."
        },
        "organizationId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Organization identifier responsible for the channel."
        },
        "names" : {
          "type" : "array",
          "description" : "Localized list of service channel names.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "descriptions" : {
          "type" : "array",
          "description" : "List of localized service channel descriptions.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "formIdentifier" : {
          "type" : "array",
          "description" : "List of localized form identifier. One per language.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "formReceiver" : {
          "type" : "array",
          "description" : "List of localized form receiver. One per language.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "deliveryAddress" : {
          "description" : "Form delivery address.",
          "$ref" : "#/definitions/Address"
        },
        "channelUrls" : {
          "type" : "array",
          "description" : "List of localized channel urls.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "attachments" : {
          "type" : "array",
          "description" : "List of attachments.",
          "items" : {
            "$ref" : "#/definitions/ServiceChannelAttachment"
          }
        },
        "supportPhones" : {
          "type" : "array",
          "description" : "List of support phone numbers for the service channel.",
          "items" : {
            "$ref" : "#/definitions/Phone"
          }
        },
        "supportEmails" : {
          "type" : "array",
          "description" : "List of support email addresses for the service channel.",
          "items" : {
            "$ref" : "#/definitions/Email"
          }
        },
        "languages" : {
          "type" : "array",
          "description" : "List of languages the service channel is available in (two letter language code).",
          "items" : {
            "type" : "string"
          }
        },
        "webPages" : {
          "type" : "array",
          "description" : "List of service channel web pages.",
          "items" : {
            "$ref" : "#/definitions/WebPage"
          }
        },
        "serviceHours" : {
          "type" : "array",
          "description" : "List of service channel service hours.",
          "items" : {
            "$ref" : "#/definitions/ServiceHour"
          }
        },
        "publishingStatus" : {
          "type" : "string",
          "description" : "Service channel publishing status. Values: Draft, Published, Deleted, Modified or OldPublished."
        }
      }
    },
    "ServiceLocationServiceChannel" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Identifier for the service channel."
        },
        "organizationId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Organization identifier responsible for the channel."
        },
        "names" : {
          "type" : "array",
          "description" : "Localized list of service channel names.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "descriptions" : {
          "type" : "array",
          "description" : "List of localized service channel descriptions.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "phoneNumbers" : {
          "type" : "array",
          "description" : "List of phone numbers for the service channel. Includes also fax numbers.",
          "items" : {
            "$ref" : "#/definitions/Phone"
          }
        },
        "emails" : {
          "type" : "array",
          "description" : "List email addresses for the service channel.",
          "items" : {
            "$ref" : "#/definitions/Email"
          }
        },
        "languages" : {
          "type" : "array",
          "description" : "List of languages the service channel is available in (two letter language code).",
          "items" : {
            "type" : "string"
          }
        },
        "phoneServiceCharge" : {
          "type" : "boolean",
          "description" : "Is the phone service charged for."
        },
        "webPages" : {
          "type" : "array",
          "description" : "List of service channel web pages.",
          "items" : {
            "$ref" : "#/definitions/WebPage"
          }
        },
        "areaType" : {
          "type" : "string",
          "description" : "Area type (WholeCountry, WholeCountryExceptAlandIslands, AreaType)."
        },
        "areas" : {
          "type" : "array",
          "description" : "List of service channel areas.",
          "items" : {
            "$ref" : "#/definitions/Area"
          }
        },
        "addresses" : {
          "type" : "array",
          "description" : "List of service location addresses.",
          "items" : {
            "$ref" : "#/definitions/Address"
          }
        },
        "serviceHours" : {
          "type" : "array",
          "description" : "List of service channel service hours.",
          "items" : {
            "$ref" : "#/definitions/ServiceHour"
          }
        },
        "publishingStatus" : {
          "type" : "string",
          "description" : "Service channel publishing status. Values: Draft, Published, Deleted, Modified or OldPublished."
        }
      }
    },
...
    "WebPage" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string",
          "description" : "Web page description."
        },
        "url" : {
          "type" : "string",
          "description" : "Web page url."
        },
        "language" : {
          "type" : "string",
          "description" : "Language code."
        },
        "value" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
    "Email" : {
      "type" : "object",
      "properties" : {
        "value" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        }
      }
    },
    "Phone" : {
      "type" : "object",
      "properties" : {
        "additionalInformation" : {
          "type" : "string"
        },
        "serviceChargeType" : {
          "type" : "string"
        },
        "chargeDescription" : {
          "type" : "string"
        },
        "prefixNumber" : {
          "type" : "string"
        },
        "isFinnishServiceNumber" : {
          "type" : "boolean"
        },
        "number" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
    "ServiceChannelAttachment" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        }
      }
    },
    "Address" : {
      "type" : "object",
      "properties" : {
        "latitude" : {
          "type" : "string",
          "description" : "Service location latitude coordinate."
        },
        "longitude" : {
          "type" : "string",
          "description" : "Service location longitude coordinate."
        },
        "coordinateState" : {
          "type" : "string",
          "description" : "State of coordinates. Coordinates are fetched from a service provided by Maanmittauslaitos (WFS).\r\nPossible values are: Loading, Ok, Failed, NotReceived, EmptyInputReceived, MultipleResultsReceived or WrongFormatReceived."
        },
        "type" : {
          "type" : "string",
          "description" : "Address type, Visiting or Postal."
        },
        "postOfficeBox" : {
          "type" : "array",
          "description" : "Post office box like PL 310",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "postalCode" : {
          "type" : "string",
          "description" : "Postal code, for example 00010."
        },
        "postOffice" : {
          "type" : "array",
          "description" : "List of localized Post offices, for example Helsinki, Helsingfors.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "streetAddress" : {
          "type" : "array",
          "description" : "List of localized street addresses.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "streetNumber" : {
          "type" : "string",
          "description" : "Street number for street address."
        },
        "municipality" : {
          "$ref" : "#/definitions/Municipality"
        },
        "country" : {
          "type" : "string",
          "description" : "Country code (ISO 3166-1 alpha-2), for example FI."
        },
        "additionalInformations" : {
          "type" : "array",
          "description" : "Localized list of additional information about the address.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        }
      }
    },
    "ServiceHour" : {
      "type" : "object",
      "properties" : {
        "serviceHourType" : {
          "type" : "string"
        },
        "validFrom" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date time where from this entry is valid."
        },
        "validTo" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date time to this entry is valid."
        },
        "isClosed" : {
          "type" : "boolean",
          "description" : "Set to true to present a time between the valid from and to times as closed."
        },
        "validForNow" : {
          "type" : "boolean",
          "description" : "Set to true to present that this entry is valid for now."
        },
        "additionalInformation" : {
          "type" : "array",
          "description" : "Localized list of additional information.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "openingHour" : {
          "type" : "array",
          "description" : "List of servicing hours (open and closes times).",
          "items" : {
            "$ref" : "#/definitions/DailyOpeningTime"
          }
        }
      }
    },
...
    "LocalizedValue" : {
      "type" : "object",
      "properties" : {
        "language" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
    "Municipality" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "string",
          "description" : "Municipality code (like 491 or 091)."
        },
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        }
      }
    },
    "Area" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string",
          "description" : "Type of the area (Municipality, Province, BusinessRegions, HospitalRegions)."
        },
        "code" : {
          "type" : "string",
          "description" : "Code of the area."
        },
        "name" : {
          "type" : "array",
          "description" : "Localized list of names for the area",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "municipalities" : {
          "type" : "array",
          "description" : "List of municipalities including municipality code and a localized list of municipality names.",
          "items" : {
            "$ref" : "#/definitions/Municipality"
          }
        }
      },
      "description" : "Area"
    },
...
    "InternalServerError" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "Forbidden" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "BadRequest" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
...
    "NotFound" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    }
  },
...
}